on_nano/requirements.txt
paho-mqtt==1.5.1
psutil==5.8.0
